Move unit tests to under a tests folder/package
