+++ lib/cylc/network/httpclient.py
@@ -712,48 +712,3 @@ def get_exception_from_html(html_text):
         return None
     return parser.exception_text
 
-
-if __name__ == '__main__':
-    import unittest
-
-    class TestSuiteRuntimeServiceClient(unittest.TestCase):
-        """Unit testing class to test the methods in SuiteRuntimeServiceClient
-        """
-        def test_url_compiler_https(self):
-            """Tests that the url parser works for a single url and command
-            using https"""
-            myclient = SuiteRuntimeServiceClient(
-                "test-suite", host=get_host(), port=80)
-            myclient.comms1[SuiteSrvFilesManager.KEY_COMMS_PROTOCOL] = 'https'
-            self.assertEqual(
-                'https://%s:80/test_command?apples=False&oranges=True' %
-                get_host(),
-                myclient._call_server_get_url(
-                    "test_command", apples="False", oranges="True"))
-
-        def test_compile_url_compiler_http(self):
-            """Test that the url compiler produces a http request when
-            http is specified."""
-            myclient = SuiteRuntimeServiceClient(
-                "test-suite", host=get_host(), port=80)
-            myclient.comms1[SuiteSrvFilesManager.KEY_COMMS_PROTOCOL] = 'http'
-            self.assertEqual(
-                'http://%s:80/test_command?apples=False&oranges=True' %
-                get_host(),
-                myclient._call_server_get_url(
-                    "test_command", apples="False", oranges="True"))
-
-        def test_compile_url_compiler_none_specified(self):
-            """Test that the url compiler produces a http request when
-            none is specified. This should retrieve it from the
-            global config."""
-            myclient = SuiteRuntimeServiceClient(
-                "test-suite", host=get_host(), port=80)
-            url = myclient._call_server_get_url(
-                "test_command", apples="False", oranges="True")
-            # Check that the url has had http (or https) appended
-            # to it. (If it does not start with "http*" then something
-            # has gone wrong.)
-            self.assertTrue(url.startswith("http"))
-
-    unittest.main()
